abstract logic from `Explore.tsx` into a custom hook
